pyproject.toml
[project]
name = "figquilt"
version = "0.1.10"
description = "figquilt is a small, language-agnostic CLI tool that composes multiple figures (PDF/SVG/PNG) into a single publication-ready figure, based on a simple layout file (YAML/JSON). The key function is creating a PDF by composing multiple PDFs and adding subfigure labels and minimal annotations."
readme = "README.md"
authors = [
    { name = "YY Ahn", email = "yongyeol@gmail.com" }
]
requires-python = ">=3.12"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "lxml>=6.0.2",
    "matplotlib>=3.10.7",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.6",
    "pyyaml>=6.0.3",
    "seaborn>=0.13.2",
    "watchfiles>=1.0.0",
]

[project.scripts]
figquilt = "figquilt.cli:main"

[project.urls]
Homepage = "https://github.com/yy/figquilt"
Repository = "https://github.com/yy/figquilt"

[build-system]
requires = ["uv_build>=0.8.12,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
